+++ src/BundlerMinifier/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace BundlerMinifier
+{
+    public class Constants
+    {
+        public const string VERSION = "1.0.21";
+    }
+}