+++ src/BundlerMinifierTest/AssemblyLoad.cs
@@ -1,0 +1,15 @@
+﻿using BundlerMinifier;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace BundlerMinifierTest
+{
+    [TestClass]
+    public class AssemblyLoad
+    {
+        [AssemblyInitialize]
+        public static void Initialize(TestContext context)
+        {
+            Telemetry.Enabled = false;
+        }
+    }
+}