+++ src/BundlerMinifierVsix/BundlerMinifierPackage.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
 using System.Windows.Threading;
+using BundlerMinifier;
 using BundlerMinifierVsix.Commands;
 using EnvDTE;
 using EnvDTE80;
@@ -11,14 +12,13 @@ using Microsoft.VisualStudio.Shell.Interop;
 namespace BundlerMinifierVsix
 {
     [PackageRegistration(UseManagedResourcesOnly = true)]
-    [InstalledProductRegistration("#110", "#112", Version, IconResourceID = 400)] // Info on this package for Help/About
+    [InstalledProductRegistration("#110", "#112", BundlerMinifier.Constants.VERSION, IconResourceID = 400)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
     [Guid(PackageGuids.guidBundlerPackageString)]
     [ProvideAutoLoad(UIContextGuids80.SolutionExists)]
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1650:ElementDocumentationMustBeSpelledCorrectly", Justification = "pkgdef, VS and vsixmanifest are valid VS terms")]
     public sealed class BundlerMinifierPackage : Package
     {
-        public const string Version = "1.0.21";
         public static DTE2 _dte;
         public static Dispatcher _dispatcher;
         public static Package Package;
@@ -26,11 +26,12 @@ namespace BundlerMinifierVsix
 
         protected override void Initialize()
         {
-            Logger.Initialize(this, Constants.VSIX_NAME);
-
             _dte = GetService(typeof(DTE)) as DTE2;
             _dispatcher = Dispatcher.CurrentDispatcher;
             Package = this;
+
+            Logger.Initialize(this, Constants.VSIX_NAME);
+            Telemetry.SetDeviceName(_dte.Edition);
 
             Events2 events = _dte.Events as Events2;
             _solutionEvents = events.SolutionEvents;