+++ src/BundlerMinifierVsix/Logger.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
+using BundlerMinifier;
 using Microsoft.VisualStudio.Shell.Interop;
 
 namespace BundlerMinifierVsix
@@ -41,6 +42,7 @@ namespace BundlerMinifierVsix
             if (ex != null)
             {
                 Log(ex.ToString());
+                Telemetry.TrackException(ex);
             }
         }